% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.FDF}
\alias{arrange.FDF}
\alias{arrange}
\title{Arrange rows by variables}
\usage{
\method{arrange}{FDF}(.data, ..., .by_group = FALSE, method = "radix")

arrange(.data, ...)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Comma separated list of unquoted variable names, or expressions
involving variable names. Use \code{\link[=desc]{desc()}} to sort a variable in descending order.}

\item{.by_group}{If \code{TRUE}, will sort first by grouping variable. Applies to
grouped data frames only.}

\item{method}{the method to be used: partial matches are allowed.  The
    default (\code{"auto"}) implies \code{"radix"} for short numeric vectors,
    integer vectors, logical vectors and factors.
    Otherwise, it implies \code{"shell"}.
    For details of methods \code{"shell"}, 
    \code{"quick"}, and \code{"radix"},
    see the help for \code{\link{sort}}.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Arrange rows by variables
}
\examples{
test <- FDF(datasets::iris)
test2 <- arrange(test, Sepal.Length)
test3 <- arrange(test, Species, desc(Sepal.Length))
}
