% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_tibble.FDF}
\alias{as_tibble.FDF}
\alias{as_tibble}
\title{Convert to tibble}
\usage{
\method{as_tibble}{FDF}(x, ...)

as_tibble(x, ...)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF}.}

\item{...}{Not used.}
}
\description{
Convert to tibble
}
\examples{
test <- FDF(datasets::iris)
test2 <- filter(test, Species == "setosa")
as_tibble(test2)
}
