% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdfr-package.R
\docType{package}
\name{bigdfr-package}
\alias{bigdfr}
\alias{bigdfr-package}
\title{bigdfr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
