% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDF-no-acc.R
\docType{methods}
\name{FDF-accessors}
\alias{FDF-accessors}
\alias{[,FDF-method}
\alias{[[,FDF-method}
\title{There is NO accessor for class \code{FDF}.}
\usage{
\S4method{[}{FDF}(x)

\S4method{[[}{FDF}(x)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF} object.}
}
\description{
There is NO accessor for class \code{FDF}.
}
