% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDF.R
\docType{methods}
\name{dim,FDF-method}
\alias{dim,FDF-method}
\alias{length,FDF-method}
\title{Dimension and type methods for class \code{FDF}.}
\usage{
\S4method{dim}{FDF}(x)

\S4method{length}{FDF}(x)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF}.}
}
\description{
Dimension and type methods for class \code{FDF}.
}
