% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{FDF_read}
\alias{FDF_read}
\title{Read a file}
\usage{
FDF_read(file, select = NULL, nb_parts = NULL, filter = NULL,
  backingfile = drop_ext(file), skip = 0, progress = FALSE, ...)
}
\arguments{
\item{file}{Path to file that you want to read.}

\item{select}{Indices of columns to keep (sorted). Default keeps them all.}

\item{nb_parts}{Number of parts in which to split reading (and transforming).
Parts are referring to blocks of selected columns.
Default use global option \code{bigreadr.part.size} to set a good value.}

\item{filter}{Vector used to subset the rows of each data frame.}

\item{backingfile}{Path to the file storing the data frame on disk.
Default uses \code{file} without its extension.}

\item{skip}{Number of lines to skip at the beginning of \code{file}.}

\item{progress}{Show progress? Default is \code{FALSE}.}

\item{...}{More arguments to be passed to \link[data.table:fread]{data.table::fread}.}
}
\value{
A Filebacked Data Frame of with <nrow> rows and <length(select)> columns.
}
\description{
Read a file as a Filebacked Data Frame by using package {bigreadr}.
}
