% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\docType{methods}
\name{filter,FDF-method}
\alias{filter,FDF-method}
\title{Return rows with matching conditions}
\usage{
\S4method{filter}{FDF}(.data, subset)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{subset}{Logical or integer vector to further subset \code{.data$ind_row}.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Use \code{filter()} find rows/cases where conditions are true. Unlike
base subsetting with \code{[}, rows where the condition evaluates to \code{NA} are
dropped.
}
\examples{
(test <- FDF(datasets::iris))
filter(test, 1:50)
filter(test, c(TRUE, FALSE, FALSE))
}
