% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by.FDF}
\alias{group_by.FDF}
\alias{group_by}
\title{Group by one or more variables}
\usage{
\method{group_by}{FDF}(.data, ..., add = FALSE)

group_by(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Expressions to compute

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where elements of
\code{vars} are objects representing their positions within
\code{vars}. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.

Note that except for \code{:}, \code{-} and \code{c()}, all complex expressions
are evaluated outside that context. This is to prevent accidental
matching to \code{vars} elements when you refer to variables from the
calling context.}

\item{add}{When \code{add = FALSE}, the default, \code{group_by()} will override
existing groups. To add to the existing groups, use \code{add = TRUE}.}
}
\description{
Group by one or more variables
}
\examples{
(test <- FDF(datasets::iris))
(test2 <- group_by(test, Species))
test2$groups
group_by(test, starts_with("Sepal"))$groups
group_by(test2, Sepal.Length, add = TRUE)$groups
}
