% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate.FDF}
\alias{mutate.FDF}
\alias{mutate}
\title{Create or transform variables}
\usage{
\method{mutate}{FDF}(.data, ...)

mutate(.data, ...)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Name-value pairs of expressions, each with length 1 or the same
length as the number of rows in the group (if using \code{\link[=group_by]{group_by()}}) or in the entire
input (if not using groups). The name of each argument will be the name of
a new variable, and the value will be its corresponding value.  Use a \code{NULL}
value in \code{mutate} to drop a variable.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
\code{mutate()} adds new variables and preserves existing ones;
\code{transmute()} adds new variables and drops existing ones.  Both
functions preserve the number of rows of the input.
}
\examples{
(test <- FDF(datasets::airquality))
mutate(test, Temp_Celsius = (Temp - 32) / 1.8,
       Temp_Kelvin = Temp_Celsius + 273.15)
}
