% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\docType{methods}
\name{pull.FDF}
\alias{pull.FDF}
\alias{pull}
\alias{pull,FDF-method}
\title{Pull out a single variable}
\usage{
pull.FDF(.data, var)

pull(.data, var = -1)

\S4method{pull}{FDF}(.data, var = -1)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{var}{For now, ONE index of the column you want to extract.}
}
\description{
This works like \code{[[} for local data frames, and automatically collects
before indexing for remote data tables.
}
\examples{
test <- FDF(datasets::iris)
pull(test, 1)
pull(test, 5)
}
