% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/as-tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{desc}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{desc}}, \code{\link[dplyr]{tibble}}}
}}

