% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\docType{methods}
\name{select.FDF}
\alias{select.FDF}
\alias{select}
\alias{select,FDF-method}
\title{Select/rename variables by name}
\usage{
select.FDF(.data, subset)

select(.data, ...)

\S4method{select}{FDF}(.data, subset)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{subset}{Logical, integer or character vector to further subset \code{.data$ind_col}.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
\code{select()} keeps only the variables you mention; \code{rename()}
keeps all variables.
}
\examples{
(test <- FDF(datasets::iris))
select(test, 1:4)
select(test, -5)
select(test, c(TRUE, FALSE))
select(test, c("Sepal.Length", "Sepal.Width"))
}
