% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarise.FDF}
\alias{summarise.FDF}
\alias{summarize}
\alias{summarise}
\title{Reduce multiple values down to a single value}
\usage{
\method{summarise}{FDF}(.data, ...)

summarize(.data, ...)

summarise(.data, ...)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Name-value pairs of summary functions. The name will be the
name of the variable in the result. The value should be an expression
that returns a single value like \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\value{
A tibble with group values and summarized values.
}
\description{
Create one or more scalar variables summarizing the variables of an
existing tbl. Tbls with groups created by \code{\link[=group_by]{group_by()}} will result in one
row in the output for each group.  Tbls with no groups will result in one row.
}
\examples{
test <- FDF(datasets::iris)
test \%>\%
  group_by(Species) \%>\%
  summarize(mean = mean(Sepal.Length))

# You can also get a list-column
test \%>\%
  group_by(Species) \%>\%
  summarize(range = range(Sepal.Length))
}
