% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\docType{methods}
\name{arrange.FDF}
\alias{arrange.FDF}
\alias{arrange}
\alias{arrange,FDF-method}
\title{Arrange rows by variables}
\usage{
arrange.FDF(.data, order)

arrange(.data, ...)

\S4method{arrange}{FDF}(.data, order)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{order}{Integer vector or orders for \code{.data$ind_row}.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Arrange rows by variables
}
\examples{
test <- FDF(datasets::iris)
pull(test, 3)
order <- order(pull(test, 3))
test2 <- arrange(test, order)
pull(test2, 3)
test3 <- filter(test, 1:50)
pull(test3, 2)
test4 <- arrange(test3, order(pull(test3, 2)))
pull(test4, 2)
}
