% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\docType{methods}
\name{arrange.FDF}
\alias{arrange.FDF}
\alias{arrange}
\alias{arrange,FDF-method}
\title{Arrange rows by variables}
\usage{
\method{arrange}{FDF}(.data, ...)

arrange(.data, ...)

\S4method{arrange}{FDF}(.data, ...)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Comma separated list of unquoted variable names. Use
\code{\link[=desc]{desc()}} to sort a variable in descending order.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Arrange rows by variables
}
\examples{
test <- FDF(datasets::iris)
test2 <- arrange(test, Sepal.Length)
}
