% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\docType{methods}
\name{as_tibble.FDF}
\alias{as_tibble.FDF}
\alias{as_tibble}
\alias{as_tibble,FDF-method}
\title{Convert to tibble}
\usage{
as_tibble.FDF(x, ...)

as_tibble(x, ...)

\S4method{as_tibble}{FDF}(x, ...)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF}.}

\item{...}{Not used.}
}
\description{
Convert to tibble
}
\examples{
test <- FDF(datasets::iris)
test2 <- filter(test, order(pull(test, 3)))
test3 <- filter(test2, 1:50)
as_tibble(test3)
}
