% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdfr-package.R
\docType{package}
\name{bigdfr-package}
\alias{bigdfr}
\alias{bigdfr-package}
\title{bigdfr: Operate with Data Frames Stored on Disk}
\description{
Operate with data frames stored in a binary file on disk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/bigdfr}
  \item Report bugs at \url{https://github.com/privefl/bigdfr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
