% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDF-no-acc.R
\name{FDF-accessors}
\alias{FDF-accessors}
\title{There is NO accessor for class \code{FDF}.}
\arguments{
\item{x}{A \link[=FDF-class]{FDF} object.}
}
\description{
There is NO accessor for class \code{FDF}.
}
