% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDF.R
\docType{class}
\name{FDF-class}
\alias{FDF-class}
\alias{FDF_RC}
\alias{FDF}
\title{Class FBM}
\usage{
FDF(df, backingfile = tempfile())
}
\arguments{
\item{df}{A data frame.}

\item{backingfile}{Path to the file storing the data on disk.
Default stores in the temporary directory.}
}
\description{
A reference class for storing and accessing data frames stored on disk.
}

