% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDF.R
\docType{class}
\name{FDF-class}
\alias{FDF-class}
\alias{FDF_RC}
\alias{FDF}
\title{Class FDF}
\usage{
FDF(df, backingfile = tempfile())
}
\arguments{
\item{df}{A data frame.}

\item{backingfile}{Path to the file storing the data on disk.
Default stores in the temporary directory.}
}
\description{
A reference class for storing and accessing data frames stored on disk.
}
\details{
An FDF object has many field:
\itemize{
\item \code{$address}: address of the external pointer containing the underlying
C++ object, to be used as a \code{XPtr<FDF>} in C++ code
\item \code{$extptr}: use \code{$address} instead
\item \code{$nrow}
\item \code{$ncol}
\item \code{$types}
\item \code{$backingfile}: File that stores the numeric data of the FDF
\item \code{$rds}: 'rds' file (that may not exist) in which this object is stored
\item \code{$is_saved}: whether this object stored in \code{$rds}?
\item \code{$ind_row}: Indices of rows for a subview of the FDF
\item \code{$ind_col}: Indices of columns for a subview view of the FDF
\item \code{$strings}: 2^16 strings to match with integers (between 0 and 65535)
\item \code{$nstr}: Number of unique strings already matched with an integer
\item \code{$groups}: Tibble with position indices of \code{$ind_row} for each group.
\item \code{$groups_internal}: use \code{$groups} instead
\item \code{$meta}: Meta information for factors, dates and character vectors.
}

And some methods:
\itemize{
\item \code{$save()}: Save the FDF object in \code{$rds}. Returns the FDF.
}
}

