% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{FDF_read}
\alias{FDF_read}
\title{Read a file}
\usage{
FDF_read(file, filter = NULL, backingfile = drop_ext(file), ...)
}
\arguments{
\item{file}{File to read.}

\item{filter}{Vector used to subset the rows of each data frame.}

\item{backingfile}{Path to the file storing the data frame on disk.
Default uses \code{file} without its extension.}

\item{...}{Arguments passed on to \code{bigreadr::big_fread2}
\describe{
  \item{nb_parts}{Number of parts in which to split reading (and transforming).
Parts are referring to blocks of selected columns.
Default uses \code{part_size} to set a good value.}
  \item{skip}{Number of lines to skip at the beginning of \code{file}.}
  \item{select}{Indices of columns to keep (sorted). Default keeps them all.}
  \item{progress}{Show progress? Default is \code{FALSE}.}
  \item{part_size}{Size of the parts if \code{nb_parts} is not supplied.
Default is \code{500 * 1024^2} (500 MB).}
}}
}
\value{
A Filebacked Data Frame.
}
\description{
Read a file as a Filebacked Data Frame by using package {bigreadr}.
}
