% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{ungroup.FDF}
\alias{ungroup.FDF}
\alias{ungroup}
\alias{group_by.FDF}
\alias{group_by}
\title{Group by one or more variables}
\usage{
\method{ungroup}{FDF}(x, ...)

ungroup(x, ...)

\method{group_by}{FDF}(.data, ..., add = FALSE)

group_by(.data, ..., add = FALSE)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF}.}

\item{...}{Not used.}

\item{.data}{A \link[=FDF-class]{FDF}.}

\item{add}{When \code{add = FALSE}, the default, \code{group_by()} will override
existing groups. To add to the existing groups, use \code{add = TRUE}.}
}
\description{
Group by one or more variables
}
\examples{
(test <- FDF(datasets::iris))
(test2 <- group_by(test, Species))
test2$groups
group_by(test, starts_with("Sepal"))$groups
group_by(test2, Sepal.Length, add = TRUE)$groups
}
