% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-tail.R
\name{head.FDF}
\alias{head.FDF}
\alias{head}
\alias{tail.FDF}
\alias{tail}
\title{Convert to a tibble of the n first/last elements.}
\usage{
\method{head}{FDF}(x, n = 6L, ...)

head(x, ...)

\method{tail}{FDF}(x, n = 6L, ...)

tail(x, ...)
}
\arguments{
\item{x}{A \link[=FDF-class]{FDF}.}

\item{n}{A single positive integer.}

\item{...}{Not used.}
}
\description{
Convert to a tibble of the n first/last elements.
}
\examples{
test <- FDF(datasets::iris)
head(test)
test <- FDF(datasets::iris)
tail(test)
}
