% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate.FDF}
\alias{mutate.FDF}
\alias{mutate}
\title{Add new variables}
\usage{
\method{mutate}{FDF}(.data, ...)

mutate(.data, ...)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop
a variable.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. They support \link[rlang:quasiquotation]{unquoting} and
splicing. See \code{vignette("programming")} for an introduction to
these concepts.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
\code{mutate()} adds new variables and preserves existing;
\code{transmute()} drops existing variables.
}
\examples{
(test <- FDF(datasets::airquality))
mutate(test, Temp_Celsius = (Temp - 32) / 1.8,
       Temp_Kelvin = Temp_Celsius + 273.15)
}
