% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\docType{methods}
\name{pull.FDF}
\alias{pull.FDF}
\alias{pull}
\alias{pull,FDF-method}
\title{Pull out a single variable}
\usage{
pull.FDF(.data, var = -1)

pull(.data, var = -1)

\S4method{pull}{FDF}(.data, var = -1)
}
\arguments{
\item{.data}{A \link[=FDF-class]{FDF}.}

\item{var}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right.
}

The default returns the last column (on the assumption that's the
column you've created most recently).

This argument is taken by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names and column positions).}
}
\description{
This works like \code{[[} for local data frames, and automatically collects
before indexing for remote data tables.
}
\examples{
test <- FDF(datasets::iris)
pull(test, 1)
pull(test, 5)
pull(test, Species)
}
